+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/definitions/parsers/DefMixinParser.java
@@ -76,7 +76,7 @@ public class DefMixinParser extends BaseDefParser {
 
     // Figure location.
     final URI uri = this.node.getToken().getURI();
-    final Location location = new Location(uri, this.node);
+    final Location location = new Location(uri, argument0Node);
 
     // Figure timestamp.
     final Instant timestamp = this.magikFile.getTimestamp();