+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/definitions/parsers/DefSlottedExemplarParser.java
@@ -98,7 +98,7 @@ public class DefSlottedExemplarParser extends BaseDefParser {
 
     // Figure location.
     final URI uri = this.node.getToken().getURI();
-    final Location location = new Location(uri, this.node);
+    final Location location = new Location(uri, argument0Node);
 
     // Figure timestamp.
     final Instant timestamp = this.magikFile.getTimestamp();
@@ -275,7 +275,7 @@ public class DefSlottedExemplarParser extends BaseDefParser {
       final List<ParameterDefinition> setParameters = Collections.emptyList();
       final ParameterDefinition assignmentParam =
           new ParameterDefinition(
-              location,
+              null,
               timestamp,
               moduleName,
               null,