+++ magik-typed-checks/src/test/java/nl/ramsolutions/sw/magik/typedchecks/checks/MagikTypedCheckTestBase.java
@@ -15,8 +15,6 @@ import nl.ramsolutions.sw.magik.typedchecks.MagikTypedCheck;
 /** Base class for MagikTypedCheck tests. */
 public class MagikTypedCheckTestBase {
 
-  private static final URI DEFAULT_URI = URI.create("memory://source.magik");
-
   /**
    * Run check on code.
    *
@@ -28,7 +26,8 @@ public class MagikTypedCheckTestBase {
   protected List<MagikIssue> runCheck(
       final String code, final IDefinitionKeeper definitionKeeper, final MagikTypedCheck check)
       throws IllegalArgumentException {
-    final MagikTypedFile magikFile = new MagikTypedFile(DEFAULT_URI, code, definitionKeeper);
+    final MagikTypedFile magikFile =
+        new MagikTypedFile(MagikTypedFile.DEFAULT_URI, code, definitionKeeper);
     return check.scanFileForIssues(magikFile);
   }