+++ CHANGES.md
@@ -12,6 +12,7 @@
 - Rename SwChar16VectorEvaluateInvocationCheck to SwChar16VectorEvaluateInvocationTypedCheck.
 - Make tab width configurable for LineLengthCheck.
 - Extend VariableNamingCheck to also check for a maximum of chars and make minLength/maxLength configurable.
+- Support renaming of methods, in case the type of the object the method is called on can be determined.
 - Several fixes.
 
 0.10.1 (2024-08-14)